Refactored elastic mapping + added some known mappings
